# panel-api/internal/api/openapi.yaml
openapi: 3.0.3
info:
  title: Jabali Panel API
  version: "v1"
  description: |
    Every route the Jabali Panel REST API serves. CI holds this file to the
    router in both directions: an undocumented route and a documented route
    that does not exist both fail the build. Routes whose only security is a
    browser session back the web UI and may change between releases; the
    token-reachable routes are the contract for scripts, CI and integrations.

    Three authentication paths:

    - **Browser sessions** (Kratos cookie) — interactive UI use
    - **Bearer tokens** (`Authorization: Bearer jat_…`) — programmatic access
      from scripts / CI / routers. Tokens are scoped to the user that minted
      them; admin tokens carry the same is_admin flag as the owning user.
    - **Automation tokens** (`Authorization: Jabali-HMAC …`, security scheme
      AutomationHMAC) — the server-wide `/automation/*` API for billing panels
      and fleet managers. It answers on :8443 (and, when an admin turns it on,
      on :443 too), but only to tokens an admin mints under
      `/admin/automation/tokens`.

    Every session and Bearer endpoint enforces ownership
    (`claims.UserID == resource.UserID`) OR `is_admin = true`. Endpoints under
    `/admin/` are admin-only. Automation tokens are server-wide and limited by
    their scopes instead.

    Token mint UI: tenant shell → API Tokens. Plaintext is shown once at
    creation; the panel only stores the SHA-256 hash thereafter.

...

...
servers:
...
components:
  securitySchemes:
    BearerToken:
      type: http
      scheme: bearer
      bearerFormat: jat_…
      description: |
        Per-user token minted via POST /me/api-tokens. Send as
        `Authorization: Bearer jat_<43-char base64url>`.
    KratosSession:
      type: apiKey
      in: cookie
      name: ory_kratos_session
      description: Browser cookie from /login. Same-origin only.
    AutomationHMAC:
      type: apiKey
      in: header
      name: Authorization
      description: |
        `Authorization: Jabali-HMAC kid=<token-id>, ts=<unix-seconds>, sig=<hex>`
        where `sig = hex(HMAC-SHA256(secret, METHOD + "\n" + URI + "\n" + ts + "\n" + hex(SHA-256(body))))`.
        `secret` is the 64-character hex string returned once at mint, used
        as-is. `URI` is the full request URI, path and query included (for
        example `/api/v1/automation/users?email=a%40example.com`). `ts` must be
        within 5 minutes of the server clock (at most 30 seconds ahead), and a
        signature is accepted once. Bodies are capped at 1 MiB.
...
security:
  - BearerToken: []
  - KratosSession: []

# -------------------------------------------------------------------- paths ---
paths:
...
  /mail/shares:
    get:
      tags: [Mail]
      summary: List shares
      security: [ { KratosSession: [] } ]
      responses: { "200": { description: OK } }


  # --- Mail Certificate (auto) ---
...
  /mailboxes/{mbid}/shares:
    get:
      tags: [Mailboxes]
      summary: Shares mailboxes
      security: [ { KratosSession: [] } ]
      parameters: [ { in: path, name: mbid, required: true, schema: { type: string } } ]
      responses: { "200": { description: OK } }
    post:
      tags: [Mailboxes]
      summary: Shares mailboxes
      security: [ { KratosSession: [] } ]
      parameters: [ { in: path, name: mbid, required: true, schema: { type: string } } ]
      responses: { "200": { description: OK } }
  /mailboxes/{mbid}/shares/{shareId}:
    delete:
      tags: [Mailboxes]
      summary: Delete share
      security: [ { KratosSession: [] } ]
      parameters: [ { in: path, name: mbid, required: true, schema: { type: string } }, { in: path, name: shareId, required: true, schema: { type: string } } ]
      responses: { "204": { description: OK } }
...
  /domains/{id}/ssl:
    delete:
      tags: [Ssl]
      summary: Ssl domains
      security: [ { KratosSession: [] } ]
      parameters: [ { in: path, name: id, required: true, schema: { type: string } } ]
      responses: { "204": { description: OK } }
    get:
      tags: [Ssl]
      summary: Ssl domains
      security: [ { KratosSession: [] } ]
      parameters: [ { in: path, name: id, required: true, schema: { type: string } } ]
      responses: { "200": { description: OK } }
    post:
      tags: [Ssl]
      summary: Ssl domains
      security: [ { KratosSession: [] } ]
      parameters: [ { in: path, name: id, required: true, schema: { type: string } } ]
      responses: { "200": { description: OK } }
...
  /admin/certificates/shared:
    post:
      tags: [Ssl Certificates]
      summary: Upload a shared wildcard/multi-SAN certificate
      security: [ { KratosSession: [] } ]
      responses: { "201": { description: Created } }
    get:
      tags: [Ssl Certificates]
      summary: List shared certificates
      security: [ { KratosSession: [] } ]
      responses: { "200": { description: OK } }
...
  /admin/certificates/shared/{id}:
    put:
      tags: [Ssl Certificates]
      summary: Replace (re-upload) a shared certificate
      security: [ { KratosSession: [] } ]
      parameters: [ { in: path, name: id, required: true, schema: { type: string } } ]
      responses: { "200": { description: OK } }
    delete:
      tags: [Ssl Certificates]
      summary: Delete a shared certificate
      security: [ { KratosSession: [] } ]
      parameters: [ { in: path, name: id, required: true, schema: { type: string } } ]
      responses: { "200": { description: OK } }
  /domains/{id}/ssl/shared:
    post:
      tags: [Ssl Certificates]
      summary: Attach a shared certificate to a domain
      security: [ { KratosSession: [] } ]
      parameters: [ { in: path, name: id, required: true, schema: { type: string } } ]
      responses: { "200": { description: OK } }
    delete:
      tags: [Ssl Certificates]
      summary: Detach a shared certificate from a domain
      security: [ { KratosSession: [] } ]
      parameters: [ { in: path, name: id, required: true, schema: { type: string } } ]
      responses: { "200": { description: OK } }
...